Cammbios en el backend y sistema de autenticacion en el front

+++ Backend/src/routes/token.js
@@ -5,42 +5,45 @@ const Token = require("../models/token");
 
 
 router.post('/codes/generate', async (req, res) => {
-    const { subject, generatedBy, isSingleUse, expirationDate } = req.body;
+    const { vacancy, generatedBy, contactMethod } = req.body;
+    const token = Math.floor(100000 + Math.random() * 900000).toString();
 
-    const expiresIn = Math.floor((new Date(expirationDate) - Date.now()) / 1000);
-    let token;
-    if (isSingleUse)   {
-        token = jwt.sign({ subject, generatedBy, isSingleUse, expirationDate}, 'secreto');
-    } else {
-        token = jwt.sign({ subject, generatedBy, isSingleUse, expirationDate}, 'secreto', { expiresIn });
+    if (!vacancy || !generatedBy || !contactMethod) {
+        return res.status(400).json({ message: 'Todos los campos son requeridos' });
     }
 
     try {
-        const newToken = new Token({ token });
+        const newToken = new Token({ token, vacancy, generatedBy, contactMethod });
         await newToken.save();
 
         res.json({ token });
     } catch (error) {
+        console.log(error)
         res.status(500).json({ message: 'Error al guardar el token' });
     }
 });
 
 
-router.post('/codes/verify', async (req, res) => {
+router.get('/codes/verify', async (req, res) => {
     const { token } = req.body;
+
+    if(!token) {
+        return res.status(400).json({ message: 'Token es requerido' });
+    }
 
     try {
         const existingToken = await Token.findOneAndDelete({ token });
 
         if (!existingToken) {
-            throw new Error('Token inválido');
+            res.status(404).json({ message: 'Token no encontrado' });
         }
 
-        const decoded = jwt.verify(token, 'secreto');
-        const { subject, generatedBy, isSingleUse, expirationDate, used } = decoded;
-        res.json({ subject, generatedBy, isSingleUse, expirationDate, used });
-        
+        await Token.deleteOne({ token });
+
+        res.json({ vacancy: existingToken.vacancy, generatedBy: existingToken.generatedBy, contactMethod: existingToken.contactMethod });
+
     } catch (error) {
+        console.log(error)
         res.status(401).json({ message: 'Token inválido' });
     }
 });
